Stabilize mobile chat GUI toggle
