+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/WorkspaceDetailView.swift
@@ -46,8 +46,8 @@ struct WorkspaceDetailView: View {
     @State private var isTextSheetPresented = false
     /// Drives the rename-workspace dialog launched from the picker menu, and its
     /// editable text (seeded with the current name when presented).
-    @State private var isRenamePresented = false
-    @State private var renameText = ""
+    @State var isRenamePresented = false
+    @State var renameText = ""
     /// Live pane width, used to width-cap the centered glass title pill so a long
     /// workspace name truncates instead of underlapping the toolbar buttons.
     @State private var contentWidth: CGFloat = 0
@@ -59,120 +59,42 @@ struct WorkspaceDetailView: View {
     /// Chat-mode toggle: when on (and a session exists) the detail renders
     /// the agent chat inline in place of the terminal. The toolbar button
     /// flips this; there is no cover and no Done button.
-    @State private var isChatMode = false
+    @State var isChatMode = false
     /// The session chat mode was entered on, pinned so a newer session
     /// sorting first cannot swap the conversation out from under the user
     /// mid-read. Cleared when chat mode turns off.
-    @State private var pinnedChatSessionID: String?
-    @State private var chatSessions: [ChatSessionDescriptor] = []
+    @State var pinnedChatSessionID: String?
+    @State var chatSessions: [ChatSessionDescriptor] = []
     /// Last terminal id whose cached snapshot said it had a chat session. This
     /// keeps the toolbar affordance and title spacing stable while SwiftUI
     /// swaps between terminal and chat mode, or while transport is reconnecting.
-    @State private var cachedChatToggleTerminalID: String?
+    @State var cachedChatToggleTerminalID: String?
     /// Per-session chat stores kept warm while the workspace detail is visible.
     /// The selected terminal's store runs even when terminal rendering is shown,
     /// so toggling into chat presents the already-loaded transcript instead of
     /// treating terminal mode as "no GUI data loaded."
-    @State private var chatConversationStores: [String: ChatConversationStore] = [:]
+    @State var chatConversationStores: [String: ChatConversationStore] = [:]
     /// Per-session composer drafts, surviving toggles back to the terminal.
-    @State private var chatDrafts: [String: String] = [:]
+    @State var chatDrafts: [String: String] = [:]
     /// App lifecycle phase. Returning from `.background` re-pulls the session
     /// list: pushes are best-effort and can be dropped while suspended, so on
     /// foreground we re-read the host's authoritative list rather than trust
     /// that every push arrived.
-    @Environment(\.scenePhase) private var scenePhase
+    @Environment(\.scenePhase) var scenePhase
     #endif
 
-    private var selectedTerminal: MobileTerminalPreview? {
+    var selectedTerminal: MobileTerminalPreview? {
         workspace.terminals.first { $0.id == store.selectedTerminalID } ?? workspace.terminals.first
     }
 
     /// Extra blank top padding for the terminal/chat, on top of the safe area. The
     /// grid already sits below the nav bar (in the top safe area), so this is just
     /// a hairline so the first row is not jammed against the bar's bottom edge.
-    private var terminalTopPadding: CGFloat { 4 }
+    var terminalTopPadding: CGFloat { 4 }
     /// The active browser surface for this workspace, when a browser pane is open.
     private var activeBrowser: BrowserSurfaceState? {
         browserStore.activeBrowser(for: workspace.id.rawValue)
     }
-
-    #if os(iOS)
-    private var selectedTerminalID: String? {
-        selectedTerminal?.id.rawValue
-    }
-
-    /// Chat sessions this view should render from right now. On first render
-    /// after returning to a workspace, local `@State` is empty, but the shell
-    /// store still has the last authoritative GUI-history snapshot for this
-    /// workspace. Use that immediately so the toolbar does not flicker while the
-    /// refresh task reconnects.
-    private var visibleChatSessions: [ChatSessionDescriptor] {
-        if !chatSessions.isEmpty {
-            return chatSessions
-        }
-        return store.cachedChatSessions(workspaceID: workspace.id.rawValue)
-    }
-
-    /// The chat session belonging to the currently visible tab/terminal, if
-    /// any. The toggle and the chat bind to THIS — the tab the user is
-    /// looking at — so a tab's chat never shows another tab's history, and a
-    /// tab with no agent session yields nil (its toggle is hidden). A past
-    /// agent that has since ended still matches here (its record keeps the
-    /// terminal binding), so the tab keeps showing the conversation read-only.
-    ///
-    /// This per-tab match relies on surface ids being stable across app
-    /// relaunch / session restore (cmux reuses a panel's persisted id when it
-    /// is still unique), so the session's recorded terminal id keeps matching
-    /// the live terminal.
-    private var sessionForSelectedTerminal: ChatSessionDescriptor? {
-        guard let terminalID = selectedTerminalID else { return nil }
-        return visibleChatSessions.first { $0.terminalID == terminalID }
-    }
-
-    /// The session backing the toolbar toggle. Prefer the currently selected
-    /// terminal's live match, but fall back to the last cached match while the
-    /// selected terminal is temporarily unavailable during mode transitions.
-    private var chatToggleSession: ChatSessionDescriptor? {
-        if let sessionForSelectedTerminal {
-            return sessionForSelectedTerminal
-        }
-        guard let terminalID = cachedChatToggleTerminalID else { return nil }
-        if let selectedTerminalID, selectedTerminalID != terminalID {
-            return nil
-        }
-        return visibleChatSessions.first { $0.terminalID == terminalID }
-    }
-
-    private var shouldShowChatToggle: Bool {
-        isChatMode || chatToggleSession != nil
-    }
-
-    /// The session chat mode opens: the visible tab's session, or the pinned
-    /// session while chat mode is on.
-    private var chosenChatSession: ChatSessionDescriptor? {
-        // While chat is open it is pinned to one session: return that exact
-        // session or nil if it vanished — never silently switch to another
-        // (the transcript/store can't follow that switch, so the header
-        // would claim B while the conversation stays A). nil makes the body
-        // fall back to the terminal and refreshChatSessions exit chat mode.
-        if let pinnedChatSessionID {
-            return visibleChatSessions.first { $0.id == pinnedChatSessionID }
-        }
-        return chatToggleSession
-    }
-
-    /// The session whose full chat model should stay warm while this detail is
-    /// visible. In terminal mode this is the selected terminal's session; in
-    /// chat mode it is the pinned session.
-    private var warmChatSession: ChatSessionDescriptor? {
-        chosenChatSession ?? chatToggleSession
-    }
-
-    /// The tab/terminal name for a session, for the chat header subtitle.
-    private func tabName(for session: ChatSessionDescriptor) -> String? {
-        workspace.terminals.first { $0.id.rawValue == session.terminalID }?.name
-    }
-    #endif
 
     var body: some View {
         Group {
@@ -208,246 +130,6 @@ struct WorkspaceDetailView: View {
     }
 
     #if os(iOS)
-    @ViewBuilder
-    private func chatContent(_ session: ChatSessionDescriptor) -> some View {
-        if let conversation = chatConversationStores[session.id] {
-            WorkspaceChatPane(
-                session: session,
-                conversation: conversation,
-                store: store,
-                workspaceName: workspace.name,
-                tabName: tabName(for: session),
-                draft: Binding(
-                    get: { chatDrafts[session.id] ?? "" },
-                    set: { chatDrafts[session.id] = $0 }
-                ),
-                onExitChat: {
-                    withAnimation(.snappy(duration: 0.28)) {
-                        isChatMode = false
-                    }
-                    pinnedChatSessionID = nil
-                }
-            )
-            .id(session.id)
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            .mobileChatTopScrollEdgeLayout(legacyTopPadding: terminalTopPadding)
-            .mobileTerminalNavigationChrome()
-            .toolbar {
-                ToolbarItemGroup(placement: .topBarTrailing) {
-                    chatToggleButton
-                    terminalPickerToolbarButton
-                }
-            }
-            .task(id: chatRefreshKey) { await refreshChatSessions() }
-            .workspaceRenameDialog(
-                isPresented: $isRenamePresented,
-                text: $renameText,
-                onSave: commitRenameFromDialog
-            )
-        } else {
-            Color.clear
-                .task(id: session.id) {
-                    _ = ensureChatConversationStore(for: session)
-                }
-        }
-    }
-
-    @ViewBuilder
-    private var chatToggleButton: some View {
-        if shouldShowChatToggle {
-            Button(action: toggleChatMode) {
-                Image(systemName: isChatMode
-                    ? "bubble.left.and.bubble.right.fill"
-                    : "bubble.left.and.bubble.right")
-            }
-            .accessibilityLabel(L10n.string("mobile.workspace.agentChat", defaultValue: "Agent Chat"))
-            .accessibilityIdentifier("MobileWorkspaceAgentChatButton")
-            .disabled(!isChatMode && chatToggleSession == nil)
-        }
-    }
-
-    /// Flip between the terminal and the inline agent chat, pinning/unpinning the
-    /// chosen session. Shared by the (legacy) toolbar button and the menu row.
-    private func toggleChatMode() {
-        let openingSession = !isChatMode ? chatToggleSession : nil
-        if let openingSession {
-            _ = ensureChatConversationStore(for: openingSession)
-        }
-        withAnimation(.snappy(duration: 0.28)) {
-            isChatMode.toggle()
-        }
-        pinnedChatSessionID = isChatMode ? openingSession?.id : nil
-    }
-
-    /// Identity for the session refetch: workspace, connection epoch, and a
-    /// foreground epoch. A change re-runs `.task(id:)`, which re-subscribes to
-    /// the push stream and re-pulls the authoritative session list.
-    ///
-    /// The foreground epoch flips only on `.background` (not transient
-    /// `.inactive` like control center or a banner), so a real
-    /// background-then-foreground re-pulls while momentary inactivity does not
-    /// churn the subscription. `.background` tears the stream down to save
-    /// battery; returning to the foreground re-establishes and reconciles.
-    private var chatRefreshKey: String {
-        let connected = store.connectionState == .connected ? 1 : 0
-        let foreground = scenePhase == .background ? 0 : 1
-        return "\(workspace.id.rawValue)#\(connected)#\(foreground)"
-    }
-
-    /// Identity for warming the selected chat model. Includes descriptor version
-    /// so an authoritative session-list refresh reconciles header/state metadata
-    /// even before a live conversation event arrives.
-    private var chatConversationWarmKey: String {
-        let connected = store.connectionState == .connected ? 1 : 0
-        let foreground = scenePhase == .background ? 0 : 1
-        guard let session = warmChatSession else {
-            return "\(workspace.id.rawValue)#none#\(connected)#\(foreground)"
-        }
-        return "\(workspace.id.rawValue)#\(session.id)#\(session.version)#\(connected)#\(foreground)"
-    }
-
-    /// Keeps the chat-capable session list current while this workspace is
-    /// shown, so the GUI toggle appears as soon as a coding agent becomes
-    /// active, without polling. The Mac pushes a `chat.message` frame on
-    /// every descriptor/state change (a brand-new agent emits
-    /// `descriptorChanged`); we register the push stream first, seed the
-    /// list once, then fold each subsequent frame in. Registering before
-    /// seeding plus idempotent folds means a change that races the seed
-    /// converges either way. The stream finishes when the connection drops;
-    /// `.task(id: chatRefreshKey)` re-runs this on reconnect, and cancels it
-    /// on workspace change or when the view goes away.
-    private func refreshChatSessions() async {
-        guard let source = store.makeChatEventSource() else {
-            applyChatModeFallback(canInvalidateSelection: false)
-            return
-        }
-        let reducer = ChatSessionListReducer(workspaceID: workspace.id.rawValue)
-        let stream = await source.sessionEvents()
-        // Animate the list update so the toggle eases in rather than popping
-        // when a session is found (the seed/first frame arriving over the
-        // wire is the "appears real quickly but not smooth" moment).
-        let seedOutcome: WorkspaceChatSessionRefreshOutcome
-        do {
-            seedOutcome = .authoritative(try await source.sessions(workspaceID: workspace.id.rawValue))
-        } catch {
-            seedOutcome = .unavailable
-        }
-        let nextSessions = seedOutcome.applying(to: visibleChatSessions)
-        withAnimation(.snappy(duration: 0.25)) {
-            chatSessions = nextSessions
-        }
-        if seedOutcome.canInvalidateSelection {
-            store.rememberChatSessions(nextSessions, workspaceID: workspace.id.rawValue)
-        }
-        refreshCachedChatToggleAnchor()
-        pruneCachedChatConversations()
-        if let warmChatSession {
-            _ = ensureChatConversationStore(for: warmChatSession)
-        }
-        repinToReopenedSession()
-        applyChatModeFallback(canInvalidateSelection: seedOutcome.canInvalidateSelection)
-        for await frame in stream {
-            let next = reducer.applying(frame, to: visibleChatSessions)
-            withAnimation(.snappy(duration: 0.25)) { chatSessions = next }
-            store.rememberChatSessions(next, workspaceID: workspace.id.rawValue)
-            refreshCachedChatToggleAnchor()
-            pruneCachedChatConversations()
-            if let warmChatSession {
-                _ = ensureChatConversationStore(for: warmChatSession)
-            }
-            repinToReopenedSession()
-            applyChatModeFallback(canInvalidateSelection: false)
-        }
-    }
-
-    /// Creates or updates the cached conversation store for a session. This is
-    /// called from tasks/actions, not from `body`, so the body remains a pure
-    /// projection of state.
-    @discardableResult
-    private func ensureChatConversationStore(
-        for session: ChatSessionDescriptor
-    ) -> ChatConversationStore? {
-        if let existing = chatConversationStores[session.id] {
-            existing.applyDescriptorSnapshot(session)
-            return existing
-        }
-        guard let source = store.makeChatEventSource() else { return nil }
-        let conversation = ChatConversationStore(descriptor: session, source: source)
-        chatConversationStores[session.id] = conversation
-        return conversation
-    }
-
-    /// Keeps cached stores bounded to sessions the workspace still knows about.
-    /// During transport loss the session list itself is preserved, so this does
-    /// not evict usable GUI state just because the Mac is reconnecting.
-    private func pruneCachedChatConversations() {
-        let liveIDs = Set(visibleChatSessions.map(\.id))
-        chatConversationStores = chatConversationStores.filter { liveIDs.contains($0.key) }
-    }
-
-    /// Runs the selected terminal's chat store while terminal mode is visible.
-    /// Opening chat reuses the same store, so there is only one subscription and
-    /// the transcript/history loaded in the background remains available.
-    private func runWarmChatConversation() async {
-        guard let session = warmChatSession,
-              let conversation = ensureChatConversationStore(for: session)
-        else { return }
-        await conversation.run()
-    }
-
-    /// Keeps the toolbar's chat affordance anchored to the latest session
-    /// snapshot for the selected terminal. Authoritative empty snapshots clear
-    /// the anchor; unavailable refreshes preserve `chatSessions`, so the anchor
-    /// naturally survives reconnects.
-    private func refreshCachedChatToggleAnchor() {
-        if let terminalID = sessionForSelectedTerminal?.terminalID {
-            cachedChatToggleTerminalID = terminalID
-            return
-        }
-
-        if let selectedTerminalID {
-            cachedChatToggleTerminalID = nil
-            return
-        }
-
-        guard let terminalID = cachedChatToggleTerminalID else { return }
-        if !visibleChatSessions.contains(where: { $0.terminalID == terminalID }) {
-            cachedChatToggleTerminalID = nil
-        }
-    }
-
-    /// While chat is open and pinned to a session that has ENDED, if the agent
-    /// was reopened on the same terminal (a newer, non-ended session bound to
-    /// the same terminal id), re-pin to it so the GUI becomes editable again.
-    /// Only an ended pin is switched, never a live one, so an active read is
-    /// never swapped out; `.id(session.id)` rebuilds the conversation store for
-    /// the new session.
-    private func repinToReopenedSession() {
-        guard isChatMode,
-              let pinnedID = pinnedChatSessionID,
-              let pinned = visibleChatSessions.first(where: { $0.id == pinnedID }),
-              pinned.state == .ended,
-              let terminalID = pinned.terminalID else { return }
-        let live = visibleChatSessions
-            .filter { $0.terminalID == terminalID && $0.id != pinnedID && $0.state != .ended }
-            .max { ($0.lastActivityAt ?? .distantPast) < ($1.lastActivityAt ?? .distantPast) }
-        if let live {
-            pinnedChatSessionID = live.id
-        }
-    }
-
-    /// If the session backing chat mode disappeared, fall back to the
-    /// terminal rather than showing an empty chat.
-    private func applyChatModeFallback(canInvalidateSelection: Bool) {
-        guard canInvalidateSelection else { return }
-        if isChatMode, chosenChatSession == nil {
-            isChatMode = false
-            pinnedChatSessionID = nil
-        }
-    }
-    #endif
-
-    #if os(iOS)
     /// The browser pane shown when this workspace has an active browser surface.
     /// It carries its own navigation chrome, so it does not get the terminal's
     /// keyboard/safe-area handling. Closing returns to the terminal.
@@ -467,9 +149,8 @@ struct WorkspaceDetailView: View {
             ToolbarItem(placement: .principal) {
                 glassTitle(browser.title ?? workspace.name)
             }
-            ToolbarItemGroup(placement: .topBarTrailing) {
-                chatToggleButton
-                terminalPickerToolbarButton
+            ToolbarItem(placement: .topBarTrailing) {
+                toolbarTrailingCluster
             }
         }
         .task(id: chatRefreshKey) { await refreshChatSessions() }
@@ -619,9 +300,8 @@ struct WorkspaceDetailView: View {
             ToolbarItem(placement: .principal) {
                 glassTitle(workspace.name)
             }
-            ToolbarItemGroup(placement: .topBarTrailing) {
-                chatToggleButton
-                terminalPickerToolbarButton
+            ToolbarItem(placement: .topBarTrailing) {
+                toolbarTrailingCluster
             }
             #else
             ToolbarItem {
@@ -697,7 +377,7 @@ struct WorkspaceDetailView: View {
     // chrome behavior is preserved; only keyboard-dismiss-on-open is dropped
     // because `Menu` has no will-open hook (the menu simply floats over the live
     // keyboard like any nav-bar menu).
-    private var terminalPickerToolbarButton: some View {
+    var terminalPickerToolbarButton: some View {
         Menu {
             terminalPickerMenuContent
         } label: {
@@ -1025,7 +705,7 @@ struct WorkspaceDetailView: View {
 
     /// Commit the rename dialog: forward the trimmed name to the Mac, which echoes
     /// it back via the authoritative list sync. Empty names are ignored.
-    private func commitRenameFromDialog() {
+    func commitRenameFromDialog() {
         let trimmed = renameText.trimmingCharacters(in: .whitespacesAndNewlines)
         guard !trimmed.isEmpty else { return }
         let store = store
